Removed deprecated components.
