+++ frontend/src/common/constants/metadata.ts
@@ -1,0 +1,8 @@
+/* Value for displaying pluralized metadata labels, for example, "tissues" or "diseases". */
+export enum PLURALIZED_METADATA_LABEL {
+  ASSAY = "assays",
+  CELL_TYPE = "cell types",
+  DISEASE = "diseases",
+  ORGANISM = "organisms",
+  TISSUE = "tissues",
+}
