--- frontend/src/components/Categories/index.tsx
@@ -1,38 +1,0 @@
-// App dependencies
-import { FC } from "react";
-// Core dependencies
-import { CategoryView, OnFilterFn } from "src/common/hooks/useCategoryFilter";
-import { Wrapper } from "./style";
-
-interface Props {
-  categories: CategoryView[];
-  onFilter: OnFilterFn;
-}
-
-/**
- * Render categories, category values and their counts. Handle click on category values.
- * TODO(cc) remove with updated datasets page.
- * @param categories - List of categories to display.
- * @param onFilter - Function to invoke on select of category value.
- */
-const Categories: FC<Props> = ({ categories, onFilter }) => {
-  return (
-    <Wrapper>
-      {categories.map((category: CategoryView) => (
-        <div key={category.key}>
-          <b>{category.label}</b>
-          {category.values.map((value) => (
-            <div
-              key={value.key}
-              onClick={() => onFilter(category.key, value.key)}
-            >
-              {value.selected ? "\u26A1" : null} {value.key} {value.count}
-            </div>
-          ))}
-        </div>
-      ))}
-    </Wrapper>
-  );
-};
-
-export default Categories;
