--- frontend/src/components/Categories/style.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-// TODO(cc) rules below are temporary and are to be updated according to mox.
-export const Wrapper = styled.div`
-  box-shadow: inset -1px 0px 0px rgba(16, 22, 26, 0.15);
-  display: flex;
-  flex-direction: column;
-  width: 240px;
-`;
