+++ frontend/src/components/Collections/components/Grid/components/Popover/index.tsx
@@ -1,4 +1,4 @@
-import { PLURALIZED_METADATA_LABEL } from "src/components/common/Filter/common/entities";
+import { PLURALIZED_METADATA_LABEL } from "src/common/constants/metadata";
 import NTag from "src/components/common/Grid/components/NTag";
 import { FieldValues } from "src/components/common/Grid/components/NTag/style";
 import { LeftAlignedDetailsCell } from "../Row/common/style";
