+++ frontend/src/components/Collections/components/Grid/components/Row/CollectionRow/index.tsx
@@ -3,6 +3,7 @@ import loadable from "@loadable/component";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { FC } from "react";
+import { PLURALIZED_METADATA_LABEL } from "src/common/constants/metadata";
 import { ROUTES } from "src/common/constants/routes";
 import { ACCESS_TYPE, VISIBILITY_TYPE } from "src/common/entities";
 import {
@@ -10,7 +11,6 @@ import {
   useCreateRevision,
 } from "src/common/queries/collections";
 import { isTombstonedCollection } from "src/common/utils/typeGuards";
-import { PLURALIZED_METADATA_LABEL } from "src/components/common/Filter/common/entities";
 import { aggregateDatasetsMetadata } from "../../../common/utils";
 import {
   LeftAlignedDetailsCell,
